chore: keep updating formatting to increase readability
This commit also contains some updates to old code that no longer makes
sense and removes the dedicated namespace for the `StateMutability` and
adds it to the abi namespace. There wasn't a need for that seperation
since they belong together and now it's also reflected in the codebase.

More commits like this should follow before working on new features for
zabi.