chore: update to latest zig master
Due to the latest writergate PR in zig alot of the interfaces and
methods have changed in zig that had to interact with the generic
implementation of the writer/readers. This commit updates the codebase
to be able to be compiled with the latest version of zig. More changes
will follow soon.