clients: update json rpc clients
This commit removes the generic clients implementation and moves
everything to be a `Provider` and `Wallet` interface.

This way we can expand the amount of providers whislt using the same
interface for all of them making things more defined.

Removed both the ens and superchain dedicated clients and added them to
the interfaces above mentioned.