clients: adapt to the new reader and writer interface
With the changes recently in zig the http and tls impl were changes
breaking some of our existing clients. This commit address those changes
and in particular updates the `WebSocketClient` to the new changes as
well.

More changes are comming since recently many breaking changes have been
happening in zig.